Use convention for get-private macros
GObject in C have a convention of using PREFIX_SUFFIX_GET_PRIVATE macros to
the private data of a GObject. PREFIX_SUFFIX_PRIVATE is also misleading, making
someone think of it as a macro for casting to a private type.
Use PREFIX_SUFFIX_GET_PRIVATE, instead of PREFIX_SUFFIX_PRIVATE.